Creativity pack: move utilities specific to Creativity pack into their own folder
