+++ src/main/java/thePackmaster/cardmodifiers/creativitypack/AccumulativeDamageModifier.java
@@ -4,7 +4,7 @@ import basemod.abstracts.AbstractCardModifier;
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
-import thePackmaster.util.JediUtil;
+import thePackmaster.util.creativitypack.JediUtil;
 
 public class AccumulativeDamageModifier extends AbstractCardModifier {
