+++ src/main/java/thePackmaster/cards/creativitypack/StrikingStrike.java
@@ -8,8 +8,7 @@ import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import thePackmaster.actions.FlexibleDiscoveryAction;
-import thePackmaster.cards.AbstractPackmasterCard;
-import thePackmaster.util.JediUtil;
+import thePackmaster.util.creativitypack.JediUtil;
 
 import static thePackmaster.SpireAnniversary5Mod.makeID;
